Fixed breakage of `brf-undo` on Emacs < 26 introduced by migration from `cl` to `cl-lib`.
The deprecation of `cl` and migration to `cl-lib` mandated by Emacs 27, unfortunately broke `brf-undo` on Emacs < 26.

It seems prior to Emacs 26, `caddr` & friends were only defined  in `cl` & `cl-lib`.
So switching to `cl-lib` meant that `caddr` only existed as `cl-caddr`.

This association with Common Lisp is a bit bizarre, as `caddr` /et al/ have existed since the earliest Lisps and indeed
are mentioned in the Lisp 1.5 manual from 1962.